--- src/main/java/Pages/Demoqa/Page_Index.java
@@ -1,20 +1,0 @@
-package Pages.Demoqa;
-
-import Helpers.Helpers;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-
-public class Page_Index {
-
-    public WebDriver driver;
-    public Helpers helpers;
-
-    public Page_Index(WebDriver driver){
-        this.driver = driver;
-        helpers = new Helpers(driver);
-    }
-
-
-
-    public By card_body_button = By.xpath("//div[@class='card_body']");
-}
