--- src/main/java/Pages/Demoqa/Page_broken.java
@@ -1,4 +1,0 @@
-package Pages.Demoqa;
-
-public class Page_broken {
-}
