--- src/main/java/Pages/Demoqa/Page_code500.java
@@ -1,4 +1,0 @@
-package Pages.Demoqa;
-
-public class Page_code500 {
-}
