--- src/main/java/Pages/Demoqa/Page_elements.java
@@ -1,4 +1,0 @@
-package Pages.Demoqa;
-
-public class Page_elements {
-}
