--- src/test/java/Test/Test_no_POM.java
@@ -1,22 +1,0 @@
-package Test;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.testng.annotations.Test;
-
-public class Test_no_POM {
-
-    public WebDriver driver;
-
-    @Test
-    public void PrimeraPrueba(){
-
-        WebDriverManager.chromedriver().setup();
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--disable-gpu");
-        driver =new ChromeDriver(options);
-        driver.get("https://demoqa.com/");
-    }
-}
